Owners can remove additional services and equipments from entity if it is not reserved. README fixed.

+++ src/main/java/com/isa/cottages/Model/Boat.java
@@ -55,7 +55,7 @@ public class Boat implements Serializable {
     private List<String> imageUrl;
 
     @Column
-    private Long capacity;
+    private Integer capacity;
 
     @Column
     private String rules;
